third_party/blink/web_tests/wpt_internal/environment-integrity/navigator-environment-integrity.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(() => {
    assert_idl_attribute(navigator, 'getEnvironmentIntegrity');
  }, 'The getEnvironmentIntegrity method should exist on the navigator object');

  promise_test(async () => {
    try {
      await navigator.getEnvironmentIntegrity('contentBinding');
      assert_unreached();
    } catch (exception) {
      assert_equals(exception.name, 'NotSupportedError');
    }
  }, 'The getEnvironmentIntegrity method throws an exception');
</script>
